refactor(markdown): land the writers' public types in morphir.langkit.markdown
ScalatagsCompiler and KyoUiCompiler declared their own packages, so compiling
a document meant importing morphir.langkit.markdown for the parser and
morphir.langkit.markdown.scalatags for the writer. That is the same "two
imports to use one library" the core module's convention already rejects; the
convention now covers all three artifacts rather than just the one.

Both writers declare `package morphir.langkit.markdown`. Their module
directories -- scalatags/ and kyo/ui/ -- still shape the artifact coordinate,
which is all they were ever for. Nothing needed hiding to make this work:
each writer's public surface is its `given Compiler[...]` instance and one
entry point, and its only implementation detail, the block separator, was
already private to the object. The internal/export machinery stays available
for when a writer grows something worth hiding.

This makes morphir.langkit.markdown a split package: three published artifacts
contribute types to it. A classpath accepts that, and it is what makes the
single import work. A Java module path would not -- JPMS forbids two modules
exporting one package, and OSGi agrees. Nothing here targets either, so the
trade is recorded in CONTRIBUTING.md and intent 0033 rather than left to be
discovered at a consumer's integration.

One trap is documented alongside it. Moving a source file between these
modules leaves incremental compilation stale, because the package does not
move with the file. Scala Native shows it worst: two different compiler
crashes, "Duplicate primitive method ==" and a cyclic reference, neither of
which points at the cause. `mill clean` on the module resolves it, and the
sources compile unchanged.

11 + 10 tests pass on JVM, Scala.js and Scala Native; morphir/ui passes on
Scala.js and Wasm.
