+++ morphir/ui/src/morphir/ui/KnowledgeBrowserView.scala
@@ -2,8 +2,7 @@ package morphir.ui
 
 import kyo.*
 import kyo.UI.*
-import morphir.langkit.markdown.Parser
-import morphir.langkit.markdown.kyoui.KyoUiCompiler
+import morphir.langkit.markdown.{KyoUiCompiler, Parser}
 import morphir.ui.services.*
 
 object KnowledgeBrowserView:
